Move RestAPIClient types and interfaces into <root>/types/rest-api-client.types.ts
