+++ packages/calypso-e2e/src/index.ts
@@ -10,6 +10,7 @@ export * from './secrets';
 export * from './email-client';
 export * from './totp-client';
 export * from './rest-api-client';
+export * from './types';
 
 export type { TracksEvent, TracksEventProperties } from './types';
 export type { TestFile } from './media-helper';
