+++ packages/calypso-e2e/src/types/index.ts
@@ -1,0 +1,1 @@
+export * from './rest-api-client.types';
