+++ test/e2e/specs/shared/api-close-account.ts
@@ -1,5 +1,5 @@
 import { RestAPIClient } from '@automattic/calypso-e2e';
-import type { AccountClosureResponse, AccountClosureDetails } from '@automattic/calypso-e2e';
+import type { AccountDetails, AccountClosureResponse } from '@automattic/calypso-e2e';
 
 /**
  * Makes an API request to close the user account.
@@ -9,12 +9,12 @@ import type { AccountClosureResponse, AccountClosureDetails } from '@automattic/
  * function is called first.
  *
  * @param {RestAPIClient} client Client to interact with the WP REST API.
- * @param {AccountClosureDetails} accountDetails Details of the account to close.
+ * @param {AccountDetails} accountDetails Details of the account to close.
  * @returns { AccountClosureResponse} Response denoting whether deletion was successful.
  */
 export async function apiCloseAccount(
 	client: RestAPIClient,
-	accountDetails: AccountClosureDetails
+	accountDetails: AccountDetails
 ): Promise< void > {
 	const response: AccountClosureResponse = await client.closeAccount( accountDetails );
